// src/services/model/evaExamenes.ts
/**
 * PostgREST API
 * standard public schema
 *
 * OpenAPI spec version: 7.0.0 (2b61a63)
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface EvaExamenes { 
    /**
     * Note: This is a Primary Key.<pk/>
     */
    idexamen?: number;
    nombre?: string;
    descripcion?: string;
    fechainicio?: string;
    fechafin?: string;
    /**
     * Note: This is a Foreign Key to `eva_admin_usuarios.idusuario`.<fk table='eva_admin_usuarios' column='idusuario'/>
     */
    idusuario?: string;
    validacion: number;
    fecharegistro?: string;
}